Fix a major bug wherein (due to my misunderstanding of the way python passes string arguments) clearing and freeing messages/keys passed into SHA2 and HMAC-SHA2 functions caused that memory to contail NULs when the variable was accessed in python.
Revert to 32bit integer types in the SHA-256 definitions as 64-bit integer types cause digest mismatches vs. OpenSSL's SHA-256. Given that the maximum message size of SHA-256 is 2^64-1, it seems that it should use 64bit integers for the key if not the context, but until I have time to trace all the shifting intricacies and verify vs. NIST test vectors, not gonna mess with it further.

These were both discovered due to the hash output not matching that of Crypto.Hash

Use SHA*_DIGEST_LEN instead of len when zeroing memory for the digest value

Remove usage of Crypto.Hash.{HMAC,SHA256} in utils.py, further reducing dependence on OpenSSL.